% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean.R
\name{clean}
\alias{clean}
\title{A convenient function for cleaning an index column}
\usage{
clean(.tbl_time, period = "yearly", side = "end")
}
\arguments{
\item{.tbl_time}{A \code{tbl_time} object.}

\item{period}{The period to use for rounding. Specified like in \code{\link[=collapse_by]{collapse_by()}}.}

\item{side}{Either "start" to round down or "end" to round up.}
}
\description{
Round the index column of your \code{tbl_time} object to a specified period.
Round down with \code{side = "start"} and round up with \code{side = "end"}.
}
\details{
\code{clean()} uses modified versions of \code{\link[lubridate:ceiling_date]{lubridate::ceiling_date()}} and
\code{\link[lubridate:floor_date]{lubridate::floor_date()}} to perform the rounding. They are modified to
work with other index classes like \code{yearmon} and \code{hms}.

\code{clean()} is most useful after a call to \code{\link[=collapse_by]{collapse_by()}} but before a
\code{\link[dplyr:group_by]{dplyr::group_by()}}.

For lower level use access to the rounding functions, use \code{\link[=ceiling_index]{ceiling_index()}}
and \code{\link[=floor_index]{floor_index()}} directly.
}
\examples{

data(FB)

FB \%>\%
  as_tbl_time(date) \%>\%
  collapse_by("2 month") \%>\%
  clean("month") \%>\%
  dplyr::group_by(date) \%>\%
  dplyr::summarise(mean_adj = mean(adjusted))

}
\seealso{
\code{\link[=ceiling_index]{ceiling_index()}}, \code{\link[=floor_index]{floor_index()}}
}
