% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/collapse_index.R
\name{collapse_by}
\alias{collapse_by}
\title{Collapse a tbl_time object's index and subsequently group by it}
\usage{
collapse_by(.tbl_time, period = "yearly", start_date = NULL, side = "end",
  ...)
}
\description{
\code{collapse_by()} is a combination of the common idiom of \code{collapse_index()}
followed by a \code{group_by()} to group on that collapsed index.
}
\details{
\code{collapse_by()} always adds to the existing \code{dplyr} groups on your \code{tbl_time}
object. Specifically it sets \code{add = TRUE} in \code{\link[dplyr:group_by]{dplyr::group_by()}}.
}
\examples{

data(FB)
FB_time <- as_tbl_time(FB, date)

collapse_by(FB_time, "month")

}
