% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/collapse_index.R
\name{collapse_index}
\alias{collapse_index}
\title{Collapse an index vector so that all observations in an interval share the
same date}
\usage{
collapse_index(index, period = "yearly", start_date = NULL, side = "end",
  ...)
}
\arguments{
\item{index}{An index vector.}

\item{period}{A formula or character specification used for time-based grouping.

If a formula, e.g. \code{1~year}, the formula is split and parsed to form the
grouping period. The \code{period} argument accepts a formula of the form
\code{multiple ~ period} allowing for flexible period grouping.
The following are examples:
\itemize{
\item 1 Year: \code{1~y}
\item 3 Months: \code{3~m}
\item 90 Days: \code{90~d}
}

Note that while shorthand is used above, an attempt is made to recognize
more explicit period names such as:
\itemize{
\item 2 Year: \code{2~year} / \code{2~years} / \code{2~yearly}
}

The \code{period} argument also accepts characters that are converted to their
corresponding periods. The following are accepted:
\itemize{
\item \code{"yearly"} or \code{"y"}
\item \code{"quarterly"} or \code{"q"}
\item \code{"monthly"} or \code{"m"}
\item \code{"weekly"} or \code{"w"}
\item \code{"daily"} or \code{"d"}
\item \code{"hour"} or \code{"h"}
\item \code{"minute"} or \code{"M"}
\item \code{"second"} or \code{"s"}
}}

\item{start_date}{Optional argument used to specify the start date for the
first group. The default is to start at the closest period boundary
below the minimum date in the supplied index.}

\item{side}{Whether to return the date at the beginning or the end of
the new period. By default, the "end" of the period.
Use "start" to change to the start of the period.}

\item{...}{Not currently used.}
}
\description{
When \code{collapse_index()} is used, the index vector is altered
so that all dates that fall in a specified interval share a common date.
The most common use case for this is to then group on the collapsed index.
}
\details{
Because this is often used for end of period summaries, the default is to
use \code{side = "end"}. Note that this is the opposite of \code{\link[=as_period]{as_period()}} where
the default is \code{side = "start"}.
}
\examples{

# Basic functionality -------------------------------------------------------

# Facebook stock prices
data(FB)
FB <- as_tbl_time(FB, date)

# Collapse to weekly dates
dplyr::mutate(FB, date = collapse_index(date, "weekly"))

# A common workflow is the group on the new date column
# to perform a time based summary
FB \%>\%
 dplyr::mutate(date = collapse_index(date, "yearly")) \%>\%
 dplyr::group_by(date) \%>\%
 dplyr::summarise_if(is.numeric, mean)


# Grouped functionality -----------------------------------------------------

data(FANG)
FANG <- FANG \%>\%
  as_tbl_time(date) \%>\%
  dplyr::group_by(symbol)

# Collapse each group to monthly,
# calculate monthly standard deviation for each column
FANG \%>\%
  dplyr::mutate(date = collapse_index(date, "monthly")) \%>\%
  dplyr::group_by(date, add = TRUE) \%>\%
  dplyr::summarise_all(sd)


}
