% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_series.R
\name{create_series}
\alias{create_series}
\title{Create a \code{tbl_time} object with a sequence of regularly spaced dates}
\usage{
create_series(time_formula, period = "daily", class = "POSIXct",
  include_end = FALSE, tz = "UTC", as_vector = FALSE)
}
\arguments{
\item{time_formula}{A period to create the series over.
This is specified as a formula.
See the \code{Details} section of \code{\link[=time_filter]{time_filter()}} for more information.}

\item{include_end}{Whether to always include the RHS of the \code{time_formula}
even if it does not match the regularly spaced index.}

\item{tz}{Time zone of the new series.}

\item{force_class}{Either \code{"Date"} or \code{"POSIXct"}. The default is to infer
the simplest class required from the \code{period} specified, but this will
override that.}

\item{as_date_vector}{Should the series be returned as a vector instead of
a tibble?}
}
\description{
\code{\link[=create_series]{create_series()}} allows the user to quickly create a \code{tbl_time} object with
a \code{date} column populated with a sequence of dates.
}
\examples{

# Every day in 2013
create_series(~2013, "daily")

# Every other day in 2013
create_series(~2013, 2~d)

# Every quarter in 2013
create_series(~2013, 1~q)

# Daily series for 2 years
create_series(2013~2015, 1~d)

# Minute series for 2 months
create_series(2012-01~2012-02, 1~M)

# Second series for 2 minutes
create_series(2011-01-01 + 12:10:00 ~ 2011-01-01 + 12:12:00, 1~s)

# Force POSIXct class
create_series(~2013, 1~d, force_class = "POSIXct")

}
