% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/index.R
\name{expand_index}
\alias{expand_index}
\title{Expand a \code{tbl_time} index into it's time signature}
\usage{
expand_index(x)
}
\arguments{
\item{x}{stuff!!}

\item{as_name}{stuff!!}
}
\description{
Returns the index column as a tibble
}
\examples{

examples!!

}
