% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getters.R
\name{getters}
\alias{getters}
\alias{get_index_quo}
\alias{get_index_char}
\alias{get_index_col}
\alias{get_index_time_zone}
\alias{get_index_class}
\title{Getters}
\usage{
get_index_quo(.tbl_time)

get_index_char(.tbl_time)

get_index_col(.tbl_time)

get_index_time_zone(.tbl_time)

get_index_class(.tbl_time)
}
\arguments{
\item{.tbl_time}{A \code{tbl_time} object.}
}
\description{
Accessors to attributes of \code{tbl_time} objects.
}
