% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/new.R
\name{new_tbl_time}
\alias{new_tbl_time}
\title{Create a new tbl_time object}
\usage{
new_tbl_time(x, index_quo, index_time_zone, ..., subclass = NULL)
}
\arguments{
\item{x}{A tibble or data.frame}

\item{index_quo}{The quo that references the index column}

\item{index_time_zone}{The index time zone}

\item{...}{Other attributes passed through to new_tibble()}

\item{subclass}{A subclass to have as a child}
}
\description{
Often used internally by developers extending tibbletime
}
