% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/to_posixct_numeric.R
\name{posixct_numeric_to_datetime}
\alias{posixct_numeric_to_datetime}
\title{Converting a posixct numeric time back to a classed datetime}
\usage{
posixct_numeric_to_datetime(x, class = "POSIXct", ..., tz = NULL)
}
\arguments{
\item{x}{A posixct numeric vector}

\item{class}{The class to convert to}

\item{...}{Extra arguments passed on the the specific coercion function}

\item{tz}{The time zone to convert to. The default UTC is used if none is
supplied}
}
\description{
Converting a posixct numeric time back to a classed datetime
}
