% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reconstruct.R
\name{reconstruct}
\alias{reconstruct}
\title{Reconstruct an S3 class from a template}
\usage{
reconstruct(new, old)
}
\arguments{
\item{new}{Freshly created object}

\item{old}{Existing object to use as template}
}
\description{
This is an implementation of \code{sloop::reconstruct()} that users can
ignore. Once \code{sloop} is on CRAN, this function will be removed and that
version will be used. It currently must be exported for use in \code{tidyquant}.
}
