% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distinct.R, R/filter.R, R/group_by.R,
%   R/mutate.R, R/select.R
\docType{import}
\name{reexports}
\alias{reexports}
\alias{distinct}
\alias{reexports}
\alias{transmute}
\alias{reexports}
\alias{filter}
\alias{reexports}
\alias{group_by}
\alias{reexports}
\alias{mutate}
\alias{reexports}
\alias{transmute}
\alias{reexports}
\alias{select}
\title{Objects exported from other packages}
\keyword{internal}
\description{
These objects are imported from other packages. Follow the links
below to see their documentation.

\describe{
  \item{dplyr}{\code{\link[dplyr]{distinct}}, \code{\link[dplyr]{transmute}}, \code{\link[dplyr]{filter}}, \code{\link[dplyr]{group_by}}, \code{\link[dplyr]{mutate}}, \code{\link[dplyr]{transmute}}, \code{\link[dplyr]{select}}}
}}

