% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/index.R
\name{retrieve_index}
\alias{retrieve_index}
\title{Retrieve a \code{tbl_time} index}
\usage{
retrieve_index(x, as_name = FALSE)
}
\arguments{
\item{x}{stuff!!}

\item{as_name}{stuff!!}
}
\description{
Returns the index column as a tibble
}
\examples{

examples!!

}
