% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/index.R
\name{retrieve_time_zone}
\alias{retrieve_time_zone}
\title{Retrieve a \code{tbl_time} time zone}
\usage{
retrieve_time_zone(x)
}
\arguments{
\item{x}{A \code{tbl_time} object}
}
\description{
Returns the \code{time_zone} attribute as a character
}
\examples{

x <- tibble::tibble(
  date  = as.Date(c('2017-01-01', '2017-01-02')),
  value = c(1, 2)
)

x <- as_tbl_time(x, date)

retrieve_time_zone(x)

}
