% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tibbletime-package.r
\docType{package}
\name{tibbletime}
\alias{tibbletime-package}
\alias{tibbletime}
\title{tibbletime: time-aware tibbles}
\description{
Built on top of the 'tibble' package, 'tibbletime' is an extension
that allows for the creation of time aware tibbles. Some immediate
advantages of this include: the ability to perform time based subsetting
on tibbles, quickly summarising and aggregating results by time periods,
and calling functions similar in spirit to the map family from 'purrr'
on time based tibbles.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/business-science/tibbletime}
  \item Report bugs at \url{https://github.com/business-science/tibbletime/issues}
}

}
\author{
\strong{Maintainer}: Davis Vaughan \email{davis@posit.co}

Authors:
\itemize{
  \item Matt Dancho \email{mdancho@business-science.io}
}

}
