% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/time_collapse.R
\name{time_collapse}
\alias{time_collapse}
\title{Collapse a \code{tbl_time} object so that all observations in a period share the
same date}
\usage{
time_collapse(.data, period = "yearly", ...)
}
\description{
Collapse a \code{tbl_time} object so that all observations in a period share the
same date
}
