% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/time_filter.R
\name{time_filter}
\alias{time_filter}
\title{Succinctly filter a \code{tbl_time} object by date}
\usage{
time_filter(x, period)
}
\arguments{
\item{x}{stuff!!}

\item{period}{stuff!!}
}
\description{
Use a concise filtering method to filter for rows where the \code{index} falls within a date range
}
\details{
Filtering is specified using the format \code{"from,to"} where the \code{from} and \code{to} are specified as:
\itemize{
\item Year - "2013,2015"
\item Month - "2013-01,2016-06"
\item Day - "2013-01-05,2016-06-04"
\item Variations - "2013,2016-06"
}
}
\examples{

examples!!

}
