% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tmap.R
\name{tmap}
\alias{tmap}
\alias{tmap_chr}
\alias{tmap_int}
\alias{tmap_lgl}
\alias{tmap_dbl}
\alias{tmap_dfc}
\alias{tmap_dfr}
\alias{tmap_df}
\title{Apply a function to each element of a \code{tbl_time} object.}
\usage{
tmap(.x, .f, period = "yearly", name = NULL, ...)

tmap_chr(.x, .f, period = "yearly", name = NULL, ...)

tmap_int(.x, .f, period = "yearly", name = NULL, ...)

tmap_lgl(.x, .f, period = "yearly", name = NULL, ...)

tmap_dbl(.x, .f, period = "yearly", name = NULL, ...)

tmap_dfc(.x, .f, period = "yearly", name = NULL, ...)

tmap_dfr(.x, .f, period = "yearly", name = NULL, ...)

tmap_df(.x, .f, period = "yearly", name = NULL, ...)
}
\arguments{
\item{.x}{A \code{tbl_time} object}

\item{.f}{A function, formula, or atomic vector.

If a \strong{function}, it is used as is.

If a \strong{formula}, e.g. \code{~ .x + 2}, it is converted to a function. There
are three ways to refer to the arguments:
\itemize{
\item For a single argument function, use \code{.}
\item For a two argument function, use \code{.x} and \code{.y}
\item For more arguments, use \code{..1}, \code{..2}, \code{..3} etc
}

This syntax allows you to create very compact anonymous functions.

If \strong{character vector}, \strong{numeric vector}, or \strong{list}, it
is converted to an extractor function. Character vectors index by name
and numeric vectors index by position; use a list to index by position
and name at different levels. Within a list, wrap strings in \code{get_attr()}
to extract named attributes. If a component is not present, the value of
\code{.default} will be returned.}

\item{period}{A period to group the mapping by}

\item{name}{The name of the list-column generated}

\item{...}{Additional arguments passed on to \code{.f}.}
}
\value{
A \code{tbl_time} object grouped by the time interval specified. The last
available date in that interval is returned as the new date.
}
\description{
The tmap functions transform a \code{tbl_time} input by applying a function to
each column at a specified time interval.
}
\details{
These functions are similar to \code{\link[purrr:map]{purrr::map()}},
but allow the user to also perform grouped mapping over
intervals such as \code{"yearly"}, \code{"monthly"}, \code{"hourly"}, etc. For example,
if \code{"monthly"} is chosen, one could picture the \code{tbl_time} object being
split into smaller \code{tbl_time}s, one for each month, and having the function
mapped over all of the columns in each of those smaller tibbles. The results
are then recombined into one tibble, with a list-column holding the results
of the mapping over each time interval.

Groupings applied using \code{\link[dplyr:group_by]{dplyr::group_by()}} are respected.
}
