% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/collapse_index.R
\name{collapse_index}
\alias{collapse_index}
\title{Collapse a \code{tbl_time} object so that all observations in a period share the
same date}
\usage{
collapse_index(index, period = "yearly", start_date = NULL,
  where = "start", normalize = FALSE, ...)
}
\arguments{
\item{...}{Not currently used.}

\item{.data}{A \code{tbl_time} object.}

\item{as_sep_col}{Whether to keep the original index as the column \code{.date}
or to drop it.}
}
\description{
When \code{time_collapse} is used, the index of a \code{tbl_time} object is altered
so that all dates that fall in a period share a common date. This can be
useful for further groupwise calculations.
}
\details{
The date chosen as the common date for a period is always the date
at the \emph{end} of that period.

It is often useful to use \code{as_sep_col = TRUE} to keep the original dates
as well.

This function respects \code{\link[dplyr:group_by]{dplyr::group_by()}} groups.
}
\examples{

# Basic functionality -------------------------------------------------------

# Facebook stock prices
data(FB)
FB <- as_tbl_time(FB, date)

# Collapse to weekly dates
time_collapse(FB, period = "weekly")

# Collapse to every other week dates
time_collapse(FB, period = 2~w)

# Collapse to weekly dates, but keep the original too
time_collapse(FB, period = "weekly", as_sep_col = TRUE)

# Grouped functionality -----------------------------------------------------

data(FANG)
FANG <- FANG \%>\%
  as_tbl_time(date) \%>\%
  dplyr::group_by(symbol)

# Collapse each group to monthly
FANG \%>\%
  time_collapse("monthly")

# Collapse each group to every other month
FANG \%>\%
  time_collapse(2~m)

}
