% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getters.R
\name{getters}
\alias{getters}
\alias{get_index_quo}
\alias{get_index_char}
\alias{get_index_col}
\alias{get_index_time_zone}
\alias{get_index_class}
\title{Getters}
\usage{
get_index_quo(x)

get_index_char(x)

get_index_col(x)

get_index_time_zone(x)

get_index_class(x)
}
\arguments{
\item{x}{A \code{tbl_time} object.}
}
\description{
Accessors to attributes of \code{tbl_time} objects.
}
