% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compat-dplyr.R, R/util.R
\docType{import}
\name{reexports}
\alias{reexports}
\alias{filter}
\alias{\%>\%}
\title{Objects exported from other packages}
\keyword{internal}
\description{
These objects are imported from other packages. Follow the links
below to see their documentation.

\describe{
  \item{dplyr}{\code{\link[dplyr]{\%>\%}}, \code{\link[dplyr]{filter}}}
}}

