% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dplyr-arrange.R, R/dplyr-distinct.R,
%   R/dplyr-filter.R, R/dplyr-group_by.R, R/dplyr-left_join.R, R/dplyr-mutate.R,
%   R/dplyr-select.R, R/dplyr-summarise.R, R/dplyr-ungroup.R, R/tidyr-gather.R,
%   R/tidyr-nest.R, R/tidyr-unnest.R, R/utils.R
\docType{import}
\name{reexports}
\alias{reexports}
\alias{arrange}
\alias{reexports}
\alias{distinct}
\alias{reexports}
\alias{filter}
\alias{reexports}
\alias{group_by}
\alias{reexports}
\alias{left_join}
\alias{reexports}
\alias{mutate}
\alias{reexports}
\alias{transmute}
\alias{reexports}
\alias{select}
\alias{reexports}
\alias{summarise}
\alias{reexports}
\alias{summarize}
\alias{reexports}
\alias{ungroup}
\alias{reexports}
\alias{gather}
\alias{reexports}
\alias{nest}
\alias{reexports}
\alias{unnest}
\alias{reexports}
\alias{\%>\%}
\title{Objects exported from other packages}
\keyword{internal}
\description{
These objects are imported from other packages. Follow the links
below to see their documentation.

\describe{
  \item{dplyr}{\code{\link[dplyr]{arrange}}, \code{\link[dplyr]{distinct}}, \code{\link[dplyr]{filter}}, \code{\link[dplyr]{group_by}}, \code{\link[dplyr]{left_join}}, \code{\link[dplyr]{mutate}}, \code{\link[dplyr]{transmute}}, \code{\link[dplyr]{select}}, \code{\link[dplyr]{summarise}}, \code{\link[dplyr]{summarize}}, \code{\link[dplyr]{ungroup}}}

  \item{magrittr}{\code{\link[magrittr]{\%>\%}}}

  \item{tidyr}{\code{\link[tidyr]{gather}}, \code{\link[tidyr]{nest}}, \code{\link[tidyr]{unnest}}}
}}

