% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tbl_time.R
\name{tbl_time}
\alias{tbl_time}
\alias{as_tbl_time}
\title{Convert an object into a \code{tbl_time} object}
\usage{
tbl_time(x, index, ...)

as_tbl_time(x, index, ...)
}
\arguments{
\item{x}{stuff!!}

\item{index}{stuff!!}

\item{...}{Currently not used}
}
\description{
\code{tbl_time} objects have a time index that contains information about which column should be
used for time based subsetting and other time based manipulation. Otherwise, they function
as normal tibbles.
}
\examples{


}
