% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/time_nest.R
\name{time_nest}
\alias{time_nest}
\title{Nest a \code{tbl_time} to a specified time period}
\usage{
time_nest(data, ..., period = "yearly", .key = data)
}
\description{
Nest a \code{tbl_time} to a specified time period
}
