% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/time_summarise.R
\name{time_summarise}
\alias{time_summarise}
\alias{time_summarize}
\title{Summarise a \code{tbl_time} by period}
\usage{
time_summarise(.data, ..., period = "yearly")

time_summarize(.data, ..., period = "yearly")
}
\description{
\code{\link[=time_summarise]{time_summarise()}} works similarly to \code{\link[dplyr:summarise]{dplyr::summarise()}} but with the added
benefit of being able to summarise by a time period such as \code{"yearly"} or
\code{"monthly"}.
}
\details{
Groups applied using \code{\link[dplyr:group_by]{dplyr::group_by()}} are respected.
}
